Removed the custom model binder and added the Ignore package
Since I decided to use the same pattern as gitignore to allow users to
ignore the directories and files they want, instead of letting them
include them, I've removed the DotNet.Glob package and added the Ignore
package. Also, because of this, the SearchPath class and the custom
model binder is no longer needed, so they have been removed.